backend/requirements.txt
Django>=5.0
psycopg[binary]>=3.2
djangorestframework>=3.15
django-cors-headers>=4.4
djangorestframework-simplejwt>=5.3
